Removed concept of logic
It was terrible to begin with.

The key was finding a workaround for {{ content }}
vs {{ node.content }}.  

That workaround lives in content-switch.html if 
you want to know more.

The other workaround is using naming your node 
variable `page` (IE `page in paginator.posts`) so
that the same code can be used to pull out page
variable when not pagination with the same 
template.

Signed-off-by: Bret Comnes <bcomnes@gmail.com>